+++ kernel/fs/vfs.c
@@ -63,8 +63,7 @@ int vfs_mount(const char *name,vfs_node *mounting_node){
 	kernel->first_mount_point = mount_point;
 	return 0;
 }
-
-int64_t vfs_read(vfs_node *node,const void *buffer,uint64_t offset,size_t count){
+ssize_t vfs_read(vfs_node *node,const void *buffer,uint64_t offset,size_t count){
 	if(node->read){
 		return node->read(node,(void *)buffer,offset,count);
 	} else {
@@ -75,7 +74,7 @@ int64_t vfs_read(vfs_node *node,const void *buffer,uint64_t offset,size_t count)
 	}
 }
 
-int64_t vfs_write(vfs_node *node,void *buffer,uint64_t offset,size_t count){
+ssize_t vfs_write(vfs_node *node,void *buffer,uint64_t offset,size_t count){
 	if(node->write){
 		return node->write(node,(void *)buffer,offset,count);
 	} else {
